requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

absl-py==1.0.0
asgiref==3.5.0
bcrypt==3.2.0
boto3==1.21.3
botocore==1.24.3
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-heroku==0.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.0
fire==0.4.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
jmespath==0.10.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mongoengine==0.23.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python-headless==4.2.0.32
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==3.12.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
sahi==0.9.0
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.1.post1
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thop==0.0.31.post2005241907
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
whitenoise==6.0.0
yolov5==6.0.6
zipp==3.7.0
